split bufferController and streamController
