+++ src/errors.js
@@ -39,5 +39,7 @@ export const ErrorDetails = {
   // Identifier for a buffer appending error event - data: appending error description
   BUFFER_APPENDING_ERROR: 'bufferAppendingError',
   // Identifier for a buffer stalled error event
-  BUFFER_STALLED_ERROR: 'bufferStalledError'
+  BUFFER_STALLED_ERROR: 'bufferStalledError',
+  // Identifier for a buffer full error event
+  BUFFER_FULL_ERROR: 'bufferFullError'
 };
